Adds Hyrax 2.2.4 as a gem dependency
Previously Hyrax was specified in Gemfile.extra, which is used by EngineCart to
load gems into the test app. But we need Hyrax as a dependency, so it was moved
to the gemspec.

Then 'hyrax' is required at the top level of the engine to avoid missing
libraries when building the test app.

The Gemspec.extra file is kept as a placeholder, because it's handy to know
it's there and what it does.