Set batch item status in callbacks
Before, BatchItem instances had their status set to 'enqueued' after they were
passed to the the BatchItemProcessingJob.perform_later. This caused issues with
batch item status when queue adapter is set to :inline (i.e. for testing),
because the job would finish, setting the batch item status to 'completed', but
then get flipped back to 'enqueued'.

By setting the batch item status to 'enqueued' in a before_enqueue callback, we
won't clobber the 'completed' status of jobs that may be run with :inline
adapter.