Pass along Reader and Ingester options
The ingester config allows implementers to specify arbitrary options in the config file
e.g. batch_ingest.yml. These options are available via instance methods #reader_options
and #ingester_options on instances of IngestConfigType objects. However, in the BatchRunner
and in the BatchItemProcessingJob objects, where Reader and Ingeter objects are
instantiated, those options were not being utilized. This PR passes the values from
IngestTypeConfig#reader_options and IngestTypeConfig#ingester_options to the constructors
of the reader and ingester classes respectively, which then are made available via
an #options attribute reader on both classes.