Avoid NoMethodError in rescue_from block
The rescue_from block in BatchItemProcessingJob was assuming that BatchItem
object was successfully looked up, but some database errors were causing
it to be nil, which would raise a subsequent NoMethodError 'update' when
trying to call #update on nil class.

This PR adds a guard against that. Also appends a full backtrace to the error
message saved to the BatchItem when in development environment.