Uses named args for BatchItemProcessingJob#perform
We had a use case where we wanted to add additional BatchItems to a Batch during
ingest, but the jobs we used required additional params for the #perform method.
By using named params to BatchItemProcessingJob#perform, we can extend
BatchItemProcessingJob and override the #perform method with any number of
arguments we want, without having to worry about our overridden #perform method
messing up the callbacks used for managing the BatchItem instance.