Do not use 'expunged' status
This reverts a previous commit where we added 'expunged' to the list of allowed
values for BatchItem#status. This was to accomodate apps who may need to remove
and object that had been ingested.

Instead of switching status to 'expunged', this PR leaves the status
as it was, uses Solr to check for the existence of the.

Also fixes display issues in batch ingest details and summary.